feature: spacy lemmatization
